+++ CHANGELOG.md
@@ -4,11 +4,19 @@ Date format: (YYYY-MM-DD)
 ## v2.13.4 (TBA)
 ### Supported MC versions: 1.17.1, 1.17, 1.16.5, 1.15.2, 1.14.4
 
+* Shopkeepers store the yaw angle now with which they are initially spawned.
+  * When placing a shopkeeper on top of a block, the yaw angle is chosen so that the shopkeeper faces towards the player who is creating the shopkeeper. When placing a shopkeeper against the side of a block, the shopkeeper is rotated according to the direction of the targeted block side.
+  * Existing shopkeepers will have a yaw of 0, i.e. they keep facing south by default. In the future it will be possible to reposition shopkeepers and thereby also adjust the yaw of already existing shopkeepers.
+  * Data: Sign shopkeepers no longer store their sign facing. Instead, this facing is now derived from the shopkeeper's yaw. Existing sign shops, for which the shopkeeper has previously not yet stored the yaw, will automatically migrate their currently stored sign facing to the shopkeeper's yaw.
+  * Shopkeeper mobs will rotate back to their initial direction now when there is no player to look at. However, this requires a player to still be somewhat nearby, since only shopkeeper mobs with nearby players are ticked.
 * Versioning: Snapshot builds will now include the Git hash in their plugin version.
 
 **API changes:**  
 * Added PlayerInactiveEvent that can be used to react to inactive players being detected, or alter which of their shopkeepers are deleted.
 * Added User interface to represent players that the plugin knows about. However, this is not yet used throughout the API.
+* Added Shopkeeper#getYaw().
+* Shopkeeper#getLocation() will include the shopkeeper's yaw now.
+* ShopCreationData#getTargetedBlockFace() no longer determines the facing of sign shops. Instead, sign shops use the yaw of the spawn location now to derive their facing.
 * Various Javadoc additions, improvements, and fixes.
 * Deprecated DefaultShopTypes#getAdminShopType() and #ADMIN(), and added #getRegularAdminShopType() and #ADMIN_REGULAR() as alternatives.
 * Deprecated the previous constructors and factory methods of AdminShopCreationData and PlayerShopCreationData, and added corresponding alternatives that directly require an AdminShopType or PlayerShopType respectively.